Add user registration, and login
Introduced a custom RegistrationForm and related logic for user sign-up, added views and templates for registration, login, and password reset, and updated URL patterns accordingly. Enhanced login and registration templates with Bootstrap styling and Google sign-in integration. Added admin registrations for models and improved session handling for sign-in and sign-out.